refactor data fetcher with per-issue trackin, adapt tests to the new refactored model
